Made changes to BHeaderView to align with changes to PParams Adjust random generation of Tests, to meet reduced scope (Word16) of MaxBHSize. Apply suggestions from code review Co-authored-by: Alexey Kuleshevich <alexey.kuleshevich@iohk.io> Added function fromIntegralBounded.

+++ eras/shelley/test-suite/src/Test/Cardano/Ledger/Shelley/Generator/Block.hs
@@ -171,7 +171,7 @@ genBlockWithTxGen
         <*> pure (fromIntegral (m * fromIntegral maxKESIterations))
         <*> pure oCert
     let hView = makeHeaderView (bheader theBlock)
-    if bhviewBSize hView <= pp ^. ppMaxBBSizeL && bhviewHSize hView <= pp ^. ppMaxBHSizeL
+    if bhviewBSize hView <= pp ^. ppMaxBBSizeL && bhviewHSize hView <= fromIntegral (pp ^. ppMaxBHSizeL)
       then pure theBlock
       else discard
     where
